refactor: #13-Redis 폴더 분리
